backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
supabase==2.0.2
langchain==0.0.352
langchain-openai==0.0.2
openai==1.6.1
PyMuPDF==1.23.8
python-docx==1.1.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sendgrid==6.11.0
gtts==2.5.1
Pillow==10.1.0
aiofiles==23.2.1
pandas==2.1.4
numpy==1.25.2
python-dotenv==1.0.0
httpx==0.24.1 